feat: substituindo a rota app/receita.tsx por app/recipes/[id].tsx, a essa nova rota adicionando a função fetchRecipe para buscar dados da API spoonacular e atualizando o path da imagem da receita para usar o URL obtido da API spoonacular
